+++ packages/context/src/services/index.ts
@@ -1,4 +1,6 @@
 export {Config} from "./Config";
 export {LoggerService} from "./LoggerService";
 export {CoreInfoService} from "./CoreInfoService";
+export {HealthService} from "./HealthService";
+export {ApplicationLifecycleBridge} from "./ApplicationLifecycleBridge";
 export * from "./types";