+++ starters/actuator/src/adapter/FastifyActuatorAdapter.ts
@@ -1,4 +1,4 @@
-import {ActuatorAdapter, ActuatorOptions, CoreInfoService} from "@nodeboot/context";
+import {ActuatorAdapter, ActuatorOptions, CoreInfoService, HealthService} from "@nodeboot/context";
 import {GitService} from "../service/GitService";
 import {MetricsContext} from "../types";
 import {MetadataService} from "../service/MetadataService";
@@ -12,6 +12,7 @@ export class FastifyActuatorAdapter implements ActuatorAdapter {
         private readonly metadataService: MetadataService,
         private readonly configService: ConfigService,
         private readonly infoService: CoreInfoService,
+        private readonly healthService: HealthService,
     ) {}
 
     bind(_: ActuatorOptions, _instance: FastifyInstance, router: FastifyInstance): void {
@@ -106,6 +107,31 @@ export class FastifyActuatorAdapter implements ActuatorAdapter {
         router.get("/actuator/middlewares", (_, res) => {
             res.status(200);
             res.send(this.metadataService.getMiddlewares());
+        });
+
+        // health
+        router.get("/actuator/health", async (_, res) => {
+            const readiness = await this.healthService.getReadiness();
+            const liveness = await this.healthService.getLiveness();
+            res.status(200);
+            res.send({
+                readinessPath: "/actuator/health/readiness",
+                livenessPath: "/actuator/health/liveness",
+                readiness,
+                liveness,
+            });
+        });
+
+        router.get("/actuator/health/readiness", async (_, res) => {
+            const {status, payload} = await this.healthService.getReadiness();
+            res.status(status);
+            res.send(payload);
+        });
+
+        router.get("/actuator/health/liveness", async (_, res) => {
+            const {status, payload} = await this.healthService.getLiveness();
+            res.status(status);
+            res.send(payload);
         });
     }
 }