+++ packages/engine/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @nodeboot/engine
 
+## 1.2.0
+
+### Minor Changes
+
+-   Upgrade typescript to latest version 5.8.3
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/context@1.7.0
+    -   @nodeboot/error@1.2.0
+
 ## 1.1.10
 
 ### Patch Changes