+++ packages/error/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @nodeboot/error
 
+## 1.2.0
+
+### Minor Changes
+
+-   Upgrade typescript to latest version 5.8.3
+
 ## 1.1.2
 
 ### Patch Changes