+++ starters/actuator/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @nodeboot/starter-actuator
 
+## 1.2.0
+
+### Minor Changes
+
+-   Add health service and extend actuator to add health endpoints
+-   Upgrade typescript to latest version 5.8.3
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/context@1.7.0
+    -   @nodeboot/config@1.2.0
+    -   @nodeboot/engine@1.2.0
+
 ## 1.1.12
 
 ### Patch Changes