+++ starters/firebase/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @nodeboot/starter-firebase
 
+## 1.1.0
+
+### Minor Changes
+
+-   Upgrade typescript to latest version 5.8.3
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/context@1.7.0
+    -   @nodeboot/core@1.9.0
+    -   @nodeboot/config@1.2.0
+    -   @nodeboot/engine@1.2.0
+
 ## 1.0.15
 
 ### Patch Changes