+++ starters/persistence/CHANGELOG.md
@@ -1,5 +1,19 @@
 # @nodeboot/starter-persistence
 
+## 1.8.0
+
+### Minor Changes
+
+-   Upgrade typescript to latest version 5.8.3
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/context@1.7.0
+    -   @nodeboot/core@1.9.0
+    -   @nodeboot/config@1.2.0
+    -   @nodeboot/di@1.2.0
+
 ## 1.7.2
 
 ### Patch Changes