+++ tests/core/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @nodeboot/test
 
+## 1.2.0
+
+### Minor Changes
+
+-   Upgrade typescript to latest version 5.8.3
+
+### Patch Changes
+
+-   Updated dependencies
+    -   @nodeboot/context@1.7.0
+    -   @nodeboot/core@1.9.0
+    -   @nodeboot/config@1.2.0
+
 ## 1.1.15
 
 ### Patch Changes